tourism_frontend/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
altair>=5.0.0
Pillow>=10.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0